Issue #5: Started propagation of the users view.
Unfortunately the pagination gem affects the :all request on the ActiveResource class and only returns the first 6 results. I can see two ways around this:

* Go through all the pages and load everyone.
* Include pagination in the client.

Either one will be a bit annoying to implement, but sort of refractoring the server code to work with and without pagination (and if it ain't broke, don't fix it) then any other solution will just be too complex.