Issue #10 Fixed the internal server error when issuing a PUT request to server/users/:id.json
This was actually an issue with the ImageService class as it was just putting all attributes into User without checking to see if non-modifiable attributes were in a hash.

Part of the problem was that ActiveResource sends all information it receives, so when Rails receives it, it may be attempting to set some non-modifiable attributes without realising.

The route I took to get around this was to remove anything not in User.attributes_accessible (seems as attributes_protected didn't include some of the inaccessible attributes) from the hashtable parameter.